requirements.txt
hpbandster==0.7.4
jupyter==1.0.0
matplotlib==3.2.1
numpy==1.18.3
pandas==0.24.2
Pillow==6.2.0
pylint==2.4.4
tensorboard==2.2.0
test-tube==0.7.5
torch==1.3.1
torchvision==0.4.2
tqdm==4.44.1
